orchestra

Trait MapSubsystem

pub trait MapSubsystem<T> {
    type Output;

    // Required method
    fn map_subsystem(&self, sub: T) -> Self::Output;
}

A helper trait to map a subsystem to smth. else.

Required Associated Types

type Output

The output type of the mapping.

Required Methods

fn map_subsystem(&self, sub: T) -> Self::Output

Consumes a T per subsystem, and maps it to Self::Output.